use k8s python binding for api access
NOTE the change was made backward-compatible with the previous
version when there is no ~/.kube/config file by replacing the default
host (https://localhost) with http://localhost:8080.

To use with a different configuration, one needs to build the pod
image with an appropriate /root/.kube/config file or have a config
file mapped into the pod's container when it is started.